feat: JpaRepository의 ID 제네릭 타입이 엔티티 PK 타입과 맞지 않던 문제 수정(#13)

+++ src/main/java/demago/khjv2/domain/studentDetail/repository/StudentDetailsRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface StudentDetailsRepository extends JpaRepository<StudentDetails,Integer> {
+public interface StudentDetailsRepository extends JpaRepository<StudentDetails,Long> {
     Optional<StudentDetails> findByStudentId(Student studentId);
 }
